Attempt to modify current font for bold
Now when preparing the "ANSIbd" tag, attempt to obtain the
current font and generate a clone of it with the weight
changed to bold.

This is done in a "new" method, so that it applies after
the superclass construction is complete, hopefully with the
font configured.

This also is done in an eval block in case it does not always work,
with the existing approach of calling $widget->Font directly
retained in the "InitObject" method.